payment-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
aio-pika==9.3.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
